tests/wpt/tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.zerosource.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.drawImage.zerosource</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.zerosource</h1>
<p class="desc">drawImage with zero-sized source rectangle draws nothing without exception</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  const response = await fetch('/images/red.png');
  const blob = await response.blob();
  const bitmap = await createImageBitmap(blob);
  ctx.drawImage(bitmap, 10, 10, 0, 1, 0, 0, 100, 50);
  ctx.drawImage(bitmap, 10, 10, 1, 0, 0, 0, 100, 50);
  ctx.drawImage(bitmap, 10, 10, 0, 0, 0, 0, 100, 50);
  _assertPixelApprox(canvas, 50,25, 0,255,0,255, 2);

}, "drawImage with zero-sized source rectangle draws nothing without exception");
</script>
